entry summary refactory - splitting into files per component
