+++ src/components/Entry/Summary/index.js
@@ -1,45 +1,30 @@
 // @flow
-import React, { PureComponent, useState } from 'react';
+import React, { PureComponent } from 'react';
 import T from 'prop-types';
 import { partition } from 'lodash-es';
 
-import ReactHtmlParser from 'react-html-parser';
-import { XMLParser } from 'fast-xml-parser';
-
-import { connect } from 'react-redux';
-import { createSelector } from 'reselect';
-import { format } from 'url';
-import { addToast } from 'actions/creators';
-
-import Link from 'components/generic/Link';
 import GoTerms from 'components/GoTerms';
 import Description from 'components/Description';
 import Literature, {
   getLiteratureIdsFromDescription,
 } from 'components/Entry/Literature';
 import CrossReferences from 'components/Entry/CrossReferences';
-import Integration from 'components/Entry/Integration';
-import ContributingSignatures from 'components/Entry/ContributingSignatures';
 import InterProHierarchy from 'components/Entry/InterProHierarchy';
-import Tooltip from 'components/SimpleCommonComponents/Tooltip';
 import Loading from 'components/SimpleCommonComponents/Loading';
-import DropDownButton from 'components/SimpleCommonComponents/DropDownButton';
 
-import loadData from 'higherOrder/loadData';
-import getUrlFor from 'utils/url-patterns';
+import MemberDBSubtitle from './MemberDBSubtitle';
+import SidePanel from './SidePanel';
+import OverlappingEntries from './OverlappingEntries';
+import Wikipedia from './Wikipedia';
 
 import { foundationPartial } from 'styles/foundation';
 
-import ipro from 'styles/interpro-new.css';
-import theme from 'styles/theme-interpro.css';
 import ebiGlobalStyles from 'ebi-framework/css/ebi-global.css';
-import fonts from 'EBI-Icon-fonts/fonts.css';
 import local from './style.css';
 
-const f = foundationPartial(ebiGlobalStyles, fonts, theme, ipro, local);
+const f = foundationPartial(ebiGlobalStyles, local);
 
-const MAX_NUMBER_OF_OVERLAPPING_ENTRIES = 5;
-/*:: type Metadata = {
+/*:: export type Metadata = {
         accession: string,
         name: {name: string, short: ?string},
         source_database: string,
@@ -65,233 +50,6 @@ const MAX_NUMBER_OF_OVERLAPPING_ENTRIES = 5;
         }
       }
  */
-
-const MemberDBSubtitle = (
-  { metadata, dbInfo } /*: {metadata: Metadata, dbInfo: Object} */,
-) => {
-  if (
-    !metadata.source_database ||
-    metadata.source_database.toLowerCase() === 'interpro'
-  ) {
-    return null;
-  }
-  return (
-    <table className={f('light', 'table-sum')}>
-      <tbody>
-        <tr>
-          <td className={f('font-ml')} style={{ width: '200px' }}>
-            Member database
-          </td>
-          <td className={f('first-letter-cap', 'md-hlight', 'font-ml')}>
-            <Link
-              className={f('nolink')}
-              to={{
-                description: {
-                  main: { key: 'entry' },
-                  entry: { db: metadata.source_database },
-                },
-              }}
-            >
-              {dbInfo.name}{' '}
-              <Tooltip
-                title={
-                  dbInfo.description || `${dbInfo.name} (${dbInfo.version})`
-                }
-              >
-                <span
-                  className={f('font-s', 'icon', 'icon-common')}
-                  data-icon="&#xf129;"
-                />
-              </Tooltip>
-            </Link>
-          </td>
-        </tr>
-        <tr>
-          <td className={f('first-letter-cap')}>{dbInfo.name} type</td>
-          <td className={f('first-letter-cap')}>
-            {metadata.type.replace('_', ' ').toLowerCase()}
-          </td>
-        </tr>
-        {metadata.name.short && metadata.accession !== metadata.name.short && (
-          <tr>
-            <td>Short name</td>
-            <td>
-              <i className={f('shortname')}>{metadata.name.short}</i>
-            </td>
-          </tr>
-        )}
-        {metadata?.counters?.sets ? (
-          <tr>
-            <td>Set</td>
-            <td>
-              {metadata.set_info ? (
-                <Link
-                  to={{
-                    description: {
-                      main: { key: 'set' },
-                      set: {
-                        db: metadata.source_database,
-                        accession: metadata.set_info.accession,
-                      },
-                    },
-                  }}
-                >
-                  {metadata.set_info.name}
-                </Link>
-              ) : (
-                'ø'
-              )}
-            </td>
-          </tr>
-        ) : null}
-      </tbody>
-    </table>
-  );
-};
-MemberDBSubtitle.propTypes = {
-  metadata: T.object.isRequired,
-  dbInfo: T.object.isRequired,
-};
-
-const _SidePanel = ({ metadata, dbInfo, api, addToast }) => {
-  const url = getUrlFor(metadata.source_database);
-  const { protocol, hostname, port, root } = api;
-
-  const [message, setMessage] = useState('');
-  const [email, setEmail] = useState('');
-
-  const apiUrl = format({
-    protocol,
-    hostname,
-    port,
-    pathname: `${root}/mail/`,
-  });
-  const entry = `${metadata.name.name} (${metadata.accession})`;
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    const data = new FormData(event.target);
-    data.append('subject', `Add annotation, ${entry}`);
-    fetch(apiUrl, {
-      method: 'POST',
-      body: data,
-    }).then((response) => {
-      let text;
-      // eslint-disable-next-line no-magic-numbers
-      if (response.status === 200) {
-        text = 'Thanks for your feedback';
-        // eslint-disable-next-line no-magic-numbers
-      } else if (response.status === 429) {
-        text = 'Request aborted as too many requests are made within a minute';
-      } else {
-        text = 'Invalid request';
-      }
-      addToast(
-        {
-          title: text,
-          ttl: 3000,
-        },
-        'interhelp-mail',
-      );
-      setMessage('');
-      setEmail('');
-    });
-  };
-
-  const handleFields = (e) => {
-    if (e.target.name === 'message') setMessage(e.target.value);
-    else setEmail(e.target.value);
-  };
-
-  const clearFields = () => {
-    setMessage('');
-    setEmail('');
-  };
-
-  return (
-    <div className={f('medium-4', 'large-4', 'columns')}>
-      <div>
-        <Tooltip
-          title={
-            'You may suggest updates to the annotation of this entry using this form. Suggestions will be sent to ' +
-            'our curators for review and, if acceptable, will be included in the next public release of InterPro. It is ' +
-            'helpful if you can include literature references supporting your annotation suggestion.'
-          }
-        >
-          <DropDownButton
-            label="Add your annotation"
-            icon="&#xf303;"
-            extraClasses={f('annotation')}
-          >
-            <form onSubmit={handleSubmit}>
-              <label htmlFor="message">Your annotation</label>
-              <textarea
-                id="message"
-                name="message"
-                value={message}
-                onChange={handleFields}
-                rows="5"
-                required
-              />
-              <label htmlFor="from_email">Email address</label>
-              <input
-                id="from_email"
-                name="from_email"
-                type="email"
-                value={email}
-                onChange={handleFields}
-                required
-              />
-              <button className={f('button')}>Submit</button>
-              <button className={f('button')} onClick={clearFields}>
-                Clear
-              </button>
-            </form>
-          </DropDownButton>
-        </Tooltip>
-      </div>
-      {metadata.integrated && <Integration intr={metadata.integrated} />}
-      {!['interpro', 'pfam'].includes(
-        metadata.source_database.toLowerCase(),
-      ) && (
-        <section>
-          <h5>External Links</h5>
-          <ul className={f('no-bullet')}>
-            <li>
-              <Link
-                className={f('ext')}
-                target="_blank"
-                href={url && url(metadata.accession)}
-              >
-                View {metadata.accession} in{' '}
-                {(dbInfo && dbInfo.name) || metadata.source_database}
-              </Link>
-            </li>
-          </ul>
-        </section>
-      )}
-      {metadata.member_databases &&
-      Object.keys(metadata.member_databases).length ? (
-        <ContributingSignatures contr={metadata.member_databases} />
-      ) : null}
-    </div>
-  );
-};
-_SidePanel.propTypes = {
-  metadata: T.object.isRequired,
-  dbInfo: T.object.isRequired,
-  api: T.object.isRequired,
-  addToast: T.func.isRequired,
-};
-
-const mapStateToProps = createSelector(
-  (state) => state.settings.api,
-  (api) => ({
-    api,
-  }),
-);
-
-const SidePanel = connect(mapStateToProps, { addToast })(_SidePanel);
 
 const OtherSections = ({ metadata, citations: { included, extra } }) => (
   <>
@@ -334,100 +92,6 @@ OtherSections.propTypes = {
   }),
 };
 
-const OverlappingEntries = ({ metadata }) => {
-  const [showAllOverlappingEntries, setShowAllOverlappingEntries] =
-    useState(false);
-  const overlaps = metadata.overlaps_with;
-  if (!overlaps || Object.keys(overlaps).length === 0) return null;
-  if (overlaps) {
-    overlaps.sort((a, b) => {
-      if (a.type > b.type) return 1;
-      if (a.type < b.type) return -1;
-      return a.accession > b.accession ? 1 : -1;
-    });
-  }
-
-  let _overlaps = overlaps;
-  if (!showAllOverlappingEntries)
-    _overlaps = metadata.overlaps_with.slice(
-      0,
-      MAX_NUMBER_OF_OVERLAPPING_ENTRIES,
-    );
-
-  return (
-    <div className={f('margin-bottom-large')}>
-      <h4>
-        {metadata.type === 'homologous_superfamily'
-          ? 'Overlapping entries'
-          : 'Overlapping homologous superfamilies'}
-        <Tooltip
-          title="The relationship between homologous superfamilies and other InterPro entries is calculated by analysing
-          the overlap between matched sequence sets. An InterPro entry is considered related to a homologous superfamily
-          if its sequence matches overlap (i.e., the match positions fall within the homologous superfamily boundaries)
-          and either the Jaccard index (equivalent) or containment index (parent/child) of the matching sequence sets is greater than 0.75."
-        >
-          &nbsp;
-          <span
-            className={f('small', 'icon', 'icon-common', 'font-s')}
-            data-icon="&#xf129;"
-          />
-        </Tooltip>
-      </h4>
-      {_overlaps.map((ov) => (
-        <div key={ov.accession} className={f('list-items')}>
-          <interpro-type type={ov.type.replace('_', ' ')} dimension="1.2em" />
-          <Link
-            to={{
-              description: {
-                main: { key: 'entry' },
-                entry: {
-                  db: 'InterPro',
-                  accession: ov.accession,
-                },
-              },
-            }}
-          >
-            {ov.name}
-          </Link>{' '}
-          <small>({ov.accession})</small>
-        </div>
-      ))}
-      {Object.keys(metadata.overlaps_with || {}).length >
-        MAX_NUMBER_OF_OVERLAPPING_ENTRIES && (
-        <button
-          className={f('button', 'hollow', 'secondary', 'margin-bottom-none')}
-          onClick={() =>
-            setShowAllOverlappingEntries(!showAllOverlappingEntries)
-          }
-        >
-          Show{' '}
-          {showAllOverlappingEntries ? (
-            <span>
-              Less{' '}
-              <i
-                className={f('icon', 'icon-common', 'font-sm')}
-                data-icon="&#xf102;"
-              />
-            </span>
-          ) : (
-            <span>
-              More{' '}
-              <i
-                className={f('icon', 'icon-common', 'font-sm')}
-                data-icon="&#xf103;"
-              />
-            </span>
-          )}
-        </button>
-      )}
-    </div>
-  );
-};
-OverlappingEntries.propTypes = {
-  metadata: T.object.isRequired,
-  overlaps: T.arrayOf(T.object),
-};
-
 const Hierarchy = ({ hierarchy, type, accession }) =>
   hierarchy &&
   Object.keys(hierarchy).length &&
@@ -445,164 +109,6 @@ Hierarchy.propTypes = {
   type: T.string,
   accession: T.string,
 };
-
-const _wikipedia = ({ title, extract, thumbnail, data }) => {
-  if (data.loading && !data.payload) return <Loading />;
-
-  const identifiers = [];
-  const article = {};
-  const properties = ['symbol', 'name', 'image', 'width', 'caption', 'pdb'];
-
-  const result = data.payload;
-  const xmlParser = new XMLParser();
-  const json = xmlParser.parse(result.parse.parsetree['*']);
-  let parts = [];
-  let infoStatus = false;
-  if (json.root.template?.length > 0) {
-    json.root.template.forEach((obj) => {
-      const possibleTitles = [
-        'Infobox protein family',
-        'Pfam_box',
-        'Pfam box',
-        'Infobox enzyme',
-      ];
-      if (possibleTitles.includes(obj.title) && !infoStatus) {
-        parts = obj.part;
-        infoStatus = true;
-      }
-    });
-  }
-
-  parts.forEach((part) => {
-    if (part.value) {
-      if (properties.includes(part.name.toLowerCase())) {
-        article[part.name] = part.value;
-        if (part.name === 'caption') {
-          if (typeof part.value === 'object') {
-            article.caption = part.value['#text'];
-          } else {
-            article.caption = part.value;
-          }
-        }
-      } else {
-        identifiers.push(part);
-      }
-    }
-  });
-
-  const imageLink = (
-    <img src={`data:image/png;base64, ${thumbnail}`} alt="Structure" />
-  );
-
-  return (
-    <div className={f('wiki-article')}>
-      <div className={f('row', 'wiki-content')}>
-        <div className={f('medium-8', 'large-8', 'columns')}>
-          <h4>
-            <Link
-              className={f('ext')}
-              target="_blank"
-              href={`https://en.wikipedia.org/wiki/${title}`}
-            >
-              {title.replace(/_/g, ' ')}
-            </Link>{' '}
-            <div className={f('tag')}>Wikipedia</div>
-          </h4>
-          {ReactHtmlParser(extract)}
-        </div>
-        <div className={f('medium-4', 'large-4', 'columns')}>
-          <table className={f('infobox')}>
-            <tbody>
-              <tr>
-                <th colSpan="2" className={f('th-infobox')}>
-                  {article.Name ? article.Name : title.replace(/_/g, ' ')}
-                </th>
-              </tr>
-              {thumbnail ? (
-                <tr>
-                  <td colSpan="2" className={f('td-thumbnail')}>
-                    {article.image ? (
-                      <a
-                        href={`https://en.wikipedia.org/wiki/File:${article.image}`}
-                        className="image"
-                        target="_blank"
-                        rel="noopener noreferrer"
-                      >
-                        {imageLink}
-                      </a>
-                    ) : (
-                      imageLink
-                    )}
-                    <div>{article.caption}</div>
-                  </td>
-                </tr>
-              ) : null}
-              {infoStatus ? (
-                <>
-                  <tr>
-                    <th colSpan="2" className={f('th-identifier')}>
-                      Identifiers
-                    </th>
-                  </tr>
-                  <tr>
-                    <th scope="row" className={f('row-header')}>
-                      Symbol
-                    </th>
-                    <td className={f('row-data')}>
-                      {article.Symbol ? article.Symbol : title}
-                    </td>
-                  </tr>
-                  {identifiers.map((id) => {
-                    return (
-                      <tr key={id.name}>
-                        <th scope="row" className={f('row-header')}>
-                          <a
-                            href={`https://en.wikipedia.org/wiki/${id.name}`}
-                            title={id.name}
-                          >
-                            {id.name}
-                          </a>
-                        </th>
-                        <td className={f('row-data')}>{id.value}</td>
-                      </tr>
-                    );
-                  })}
-                </>
-              ) : null}
-            </tbody>
-          </table>
-        </div>
-      </div>
-    </div>
-  );
-};
-_wikipedia.propTypes = {
-  title: T.string,
-  extract: T.string,
-  thumbnail: T.string,
-  data: T.object,
-};
-
-const getWikiUrl = createSelector(
-  (state) => state.settings.wikipedia,
-  (_state, props) => props.title,
-  ({ protocol, hostname, port, root }, title) => {
-    return format({
-      protocol,
-      hostname,
-      port,
-      pathname: root,
-      query: {
-        origin: '*',
-        action: 'parse',
-        page: title,
-        format: 'json',
-        prop: 'parsetree',
-      },
-    });
-  },
-);
-const Wikipedia = loadData({ getUrl: getWikiUrl })(_wikipedia);
 
 /*:: type Props = {
     data: {
@@ -642,10 +148,8 @@ class SummaryEntry extends PureComponent /*:: <Props> */ {
           <div className={f('row')}>
             <div className={f('medium-8', 'large-8', 'columns')}>
               <MemberDBSubtitle metadata={metadata} dbInfo={dbInfo} />
-              {metadata.source_database &&
-                metadata.source_database.toLowerCase() === 'interpro' &&
-                metadata.name.short &&
-                metadata.accession !== metadata.name.short && (
+              {metadata?.source_database?.toLowerCase() === 'interpro' &&
+                metadata?.accession !== metadata?.name?.short && (
                   <p data-testid="entry-shortname">
                     Short name:&nbsp;
                     <i className={f('shortname')}>{metadata.name.short}</i>
@@ -672,7 +176,9 @@ class SummaryEntry extends PureComponent /*:: <Props> */ {
                 ) : null
               }
             </div>
-            <SidePanel metadata={metadata} dbInfo={dbInfo} />
+            <div className={f('medium-4', 'large-4', 'columns')}>
+              <SidePanel metadata={metadata} dbInfo={dbInfo} />
+            </div>
           </div>
         </section>
         <OtherSections metadata={metadata} citations={{ included, extra }} />
